♻️refactor (api): split `verifyAuth` and Custom Error classes
